feat: пока так 7

+++ ewm-main-service/src/main/java/ru/practicum/main/service/AdminCategoriesServiceImpl.java
@@ -7,6 +7,8 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.main.dao.CategoriesMainServiceRepository;
+import ru.practicum.main.dao.EventMainServiceRepository;
+import ru.practicum.main.exception.ConflictException;
 import ru.practicum.main.exception.NotFoundException;
 import ru.practicum.main.model.Categories;
 
@@ -19,6 +21,8 @@ public class AdminCategoriesServiceImpl implements AdminCategoriesService {
 
     private final CategoriesMainServiceRepository repository;
 
+    private final EventMainServiceRepository eventMainServiceRepository;
+
     @Override
     @Transactional
     public Categories createCategories(Categories categories) {
@@ -28,6 +32,10 @@ public class AdminCategoriesServiceImpl implements AdminCategoriesService {
     @Override
     @Transactional
     public void deleteCategories(long catId) {
+        if (eventMainServiceRepository.existsByCategoryId(catId)) {
+            throw new ConflictException("Нельзя удалить категорию с ивентами");
+        }
+
         boolean answer = repository.existsById(catId);
         if (answer) {
             repository.deleteById(catId);
